Remove the hosting and options dependencies from the ModelContextProtocol package
