+++ samples/QuickstartWeatherServer/QuickstartWeatherServer.csproj
@@ -13,7 +13,7 @@
     </ItemGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\..\src\ModelContextProtocol\ModelContextProtocol.csproj" />
+        <ProjectReference Include="..\..\src\ModelContextProtocol.AspNetCore\ModelContextProtocol.AspNetCore.csproj" />
     </ItemGroup>
 
 </Project>
