+++ src/ModelContextProtocol.AspNetCore/Configuration/IMcpServerBuilder.cs
