+++ src/ModelContextProtocol.AspNetCore/Configuration/McpServerHandlers.cs
