+++ src/ModelContextProtocol.AspNetCore/Configuration/McpServerServiceCollectionExtensions.cs
@@ -1,5 +1,6 @@
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Options;
+using ModelContextProtocol.AspNetCore.Configuration;
 using ModelContextProtocol.Server;
 
 namespace Microsoft.Extensions.DependencyInjection;
