+++ src/ModelContextProtocol/ModelContextProtocol.csproj
@@ -24,6 +24,7 @@
     <PackageReference Include="Microsoft.Bcl.Memory" />
     <PackageReference Include="System.Text.Json" />
     <PackageReference Include="System.Threading.Channels" />
+    <PackageReference Include="System.Diagnostics.DiagnosticSource" />
   </ItemGroup>
 
   <!-- Dependencies only needed by netstandard2.0 or net8.0 -->
@@ -34,7 +35,6 @@
   <!-- Dependencies needed by all -->
   <ItemGroup>
     <PackageReference Include="Microsoft.Extensions.AI.Abstractions" />
-    <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" />
     <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" />
     <PackageReference Include="System.Net.ServerSentEvents" />
   </ItemGroup>
