+++ src/ModelContextProtocol/Server/McpServerPromptTypeAttribute.cs
@@ -9,9 +9,8 @@ namespace ModelContextProtocol.Server;
 /// <para>
 /// This attribute is used to mark a class containing methods that should be automatically
 /// discovered and registered as <see cref="McpServerPrompt"/>s. When combined with discovery methods like
-/// <see cref="McpServerBuilderExtensions.WithPromptsFromAssembly"/>, it enables automatic registration 
-/// of prompts without explicitly listing each prompt class. The attribute is not necessary when a reference
-/// to the type is provided directly to a method like <see cref="McpServerBuilderExtensions.WithPrompts{T}"/>.
+/// WithPromptsFromAssembly, it enables automatic registration of prompts without explicitly listing each prompt class.
+/// The attribute is not necessary when a reference to the type is provided directly to a method like WithPrompts.
 /// </para>
 /// <para>
 /// Within a class marked with this attribute, individual methods that should be exposed as
