+++ src/ModelContextProtocol/Server/McpServerToolAttribute.cs
@@ -11,7 +11,7 @@ namespace ModelContextProtocol.Server;
 /// <remarks>
 /// <para>
 /// This attribute is applied to methods that should be exposed as tools in the Model Context Protocol. When a class 
-/// containing methods marked with this attribute is registered with <see cref="McpServerBuilderExtensions"/>,
+/// containing methods marked with this attribute is registered with McpServerBuilderExtensions,
 /// these methods become available as tools that can be called by MCP clients.
 /// </para>
 /// <para>
