+++ src/ModelContextProtocol/Server/McpServerToolTypeAttribute.cs
@@ -9,9 +9,8 @@ namespace ModelContextProtocol.Server;
 /// <para>
 /// This attribute is used to mark a class containing methods that should be automatically
 /// discovered and registered as <see cref="McpServerTool"/>s. When combined with discovery methods like
-/// <see cref="McpServerBuilderExtensions.WithToolsFromAssembly"/>, it enables automatic registration 
-/// of tools without explicitly listing each tool class. The attribute is not necessary when a reference
-/// to the type is provided directly to a method like <see cref="McpServerBuilderExtensions.WithTools{T}"/>.
+/// WithToolsFromAssembly, it enables automatic registration of tools without explicitly listing each tool
+/// class. The attribute is not necessary when a reference to the type is provided directly to a method like WithTools.
 /// </para>
 /// <para>
 /// Within a class marked with this attribute, individual methods that should be exposed as
